Closes #654
